+++ SDLTooSharp/Bindings/SDL2Mixer/SDLMixer.cs
@@ -7,7 +7,16 @@ namespace SDLTooSharp.Bindings.SDL2Mixer;
 public static partial class SDLMixer
 {
     private const string dllName = "SDL2_mixer";
+    
+    [DllImport(dllName, CallingConvention = CallingConvention.Cdecl, EntryPoint = "Mix_Linked_Version")]
+    private static extern IntPtr _Mix_Linked_Version();
 
+    public static SDL.SDL_version Mix_Linked_Version()
+    {
+        IntPtr version = _Mix_Linked_Version();
+        return Marshal.PtrToStructure<SDL.SDL_version>(version);
+    }
+    
     public enum MIX_InitFlags
     {
         MIX_INIT_FLAC = 0x00000001,