+++ SDLTooSharp/Bindings/SDL2/Metal.cs
@@ -1,0 +1,18 @@
+﻿using System.Runtime.InteropServices;
+
+namespace SDLTooSharp.Bindings.SDL2;
+
+public static partial class SDL
+{
+    [DllImport(dllName, CallingConvention = CallingConvention.Cdecl)]
+    public static extern IntPtr SDL_Metal_CreateView(IntPtr window);
+
+    [DllImport(dllName, CallingConvention = CallingConvention.Cdecl)]
+    public static extern void SDL_Metal_DestroyView(IntPtr metalView);
+
+    [DllImport(dllName, CallingConvention = CallingConvention.Cdecl)]
+    public static extern IntPtr SDL_Metal_GetLayer(IntPtr metalView);
+
+    [DllImport(dllName, CallingConvention = CallingConvention.Cdecl)]
+    public static extern void SDL_Metal_GetDrawableSize(IntPtr window, out int w, out int h);
+}