+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix const types in Audio.cs.
 - Fix return type of `SDL_GetSystemRAM`.
+- Fix types in `SDL_KeyboardEvent` struct.
 
 ## [0.1.1] - 2022-09-25
 ### Fixed