+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix return type of `SDL_GetSystemRAM`.
 - Fix types in `SDL_KeyboardEvent` struct.
 - Fix `SDL_JoyAxisEvent` struct.
+- Fix `SDL_PeepEvents` signature, where `events` parameter was not an array.
 
 ## [0.1.1] - 2022-09-25
 ### Fixed