+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added missing function `SDL_LoadWAV`.
-
+- Added missing function `SDL_GameControllerAddMappingsFromFile`.
 ### Fixed
 - Fix const types in Audio.cs.
 - Fix return type of `SDL_GetSystemRAM`.