+++ CHANGELOG.md
@@ -9,12 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added missing function `SDL_LoadWAV`.
 - Added missing function `SDL_GameControllerAddMappingsFromFile`.
+
 ### Fixed
 - Fix const types in Audio.cs.
 - Fix return type of `SDL_GetSystemRAM`.
 - Fix types in `SDL_KeyboardEvent` struct.
 - Fix `SDL_JoyAxisEvent` struct.
 - Fix `SDL_PeepEvents` signature, where `events` parameter was not an array.
+- Fix `SDL_GameControllerGetSensorData` signature, where data was not an array.
 
 ## [0.1.1] - 2022-09-25
 ### Fixed