+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added missing function `SDL_LoadWAV`.
 - Added missing function `SDL_GameControllerAddMappingsFromFile`.
 - Added missing function `SDL_JoystickUpdate`.
+- Added metal related functions (`metal.h`)
+
 
 ### Fixed
 - Fix const types in Audio.cs.