+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added missing function `SDL_GameControllerAddMappingsFromFile`.
 - Added missing function `SDL_JoystickUpdate`.
 - Added metal related functions (`metal.h`)
+- Added system related functions (`system.h`).
 
 
 ### Fixed