+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added metal related functions (`metal.h`)
 - Added system related functions (`system.h`).
 - Added version related function (`version.h`).
-
+- Added missing version functions to SDL_Image, SDL_Mixer and SDL_TTF.
 
 ### Fixed
 - Fix const types in Audio.cs.