+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-09-25
 ### Added
 - Added missing function `SDL_LoadWAV`.
 - Added missing function `SDL_GameControllerAddMappingsFromFile`.
@@ -35,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bindings for SDL2 Image
 - Bindings for SDL2 Ttf
 
-[Unreleased]: https://github.com/MisterIcy/SDLTooSharp/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/MisterIcy/SDLTooSharp/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/MisterIcy/SDLTooSharp/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/MisterIcy/SDLTooSharp/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/MisterIcy/SDLTooSharp/releases/tag/v0.1.0